# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.3.0",
    "description": "Go-native Rebecca panel API for dashboard, bot, and third-party client developers.\n\nAuthentication: the dashboard uses an opaque server-side session stored in the HttpOnly `rebecca_session` cookie. Session permissions are loaded from the database on every request, so permission changes take effect without logging out. Bots and external integrations should use personal `rk_...` API keys as `Authorization: Bearer <api_key>`. Legacy JWT creation at `POST /admin/token` and `POST /api/admin/token` remains available only for older Marzban-compatible clients.\n\nDo not automate username/password login for bots. API keys can be revoked independently and inherit the owner admin permissions, service scope, and limits.\n\nMutation behavior: user, service, node, config, and host mutations are transactional where applicable. Runtime changes are synchronized to nodes through gRPC and recorded for recovery."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/settings/backup/export": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Export backup",
        "description": "Export Rebecca backup in .rbbackup format on binary installations.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Panel read requires an authenticated admin. Settings mutation, backup, templates, and TODO/410 settings routes require sudo/full_access.\n\nRoute details: Only works in binary installation mode. The response is a `.rbbackup` archive with manifest and database/files depending on scope.",
        "responses": {
          "200": {
            "description": "Backup archive."
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "getSettingsBackupExport"
      }
    },
    "/api/settings/backup/import": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Import backup",
        "description": "Import a Rebecca backup archive and restore database/files according to backup scope.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Panel read requires an authenticated admin. Settings mutation, backup, templates, and TODO/410 settings routes require sudo/full_access.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.\n\nRoute details: Only works in binary installation mode. Import performs safe extraction and hard restore; use a maintenance window for production restores.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "operationId": "postSettingsBackupImport"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Rebecca admin API key or legacy JWT",
        "description": "Use a personal rk_ API key for bots and third-party integrations. Legacy JWTs from /admin/token remain accepted for older Marzban-compatible clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rebecca_session",
        "description": "Opaque HttpOnly dashboard session cookie issued by POST /api/auth/login. It is managed by the browser and is not readable by dashboard JavaScript."
      }
    },
...
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible Rebecca API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            },
            "example": {
              "id": 1,
              "status": "ok",
              "detail": "Operation completed"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error response. Most handlers return `{ \"detail\": \"...\" }`; some compatibility routes may return a plain text body when handled by net/http helpers.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "AnyValue": {
        "nullable": true,
        "example": {
          "id": 1,
          "status": "ok"
        },
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}